Fix bug in handling API errors
